+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `set_lna_gain` method
 - `valid_rx_headers` method
 - `valid_rx_packets` method
+- `last_packet_snr` method
 
 ## [0.1.0] - 2026-03-07